Add `accent_gray_status` option
After using this theme at work for a couple of weeks, I decided the
colorful statusline was definetly the most uncomfy element of it. So I
made it easy to turn it off with this option; this is what I will be
using from now on. Default behavior is, of course, unchanged; there
shall be no breaking changes!