Less highlighting, tweak gray statusline colors
The accent color usage is now even more restrained, with numbers and
user defined constants using the default foreground color. This makes it
easy to differentiate 0 from "0", for instance, and makes the accent
color pop out even more. As for the statusline, it no longer uses the
same background as cursorline, which is nice.